refactor: split archive and runtime health layers
